添加附魔界面各个模块

+++ app/src/main/java/com/mika/toramhelper/enchantment/bean/EProperty.java
@@ -35,7 +35,7 @@ public class EProperty {
         this.name = name;
     }
 
-    public float getPotential() {
+    public int getPotential() {
         return potential;
     }
